login and register templates updated
