+++ src/components/SignIn.js
@@ -3,7 +3,7 @@ import { grid } from '@material-ui/core';
 import {connect} from 'react-redux'
 import {signIn} from '../store/actions/authActions'
 import { Redirect } from 'react-router-dom'
-import {firebase}
+
